requirements.txt
pyrogram
TgCrypto
python-dotenv
hachoir
ffmpeg-python
